Big update
Remove VS projects
Full building using MinGW
Get major and minor version from gtMainSystem

+++ source/gtrD3D11/gtTextureD3D11.cpp
@@ -22,7 +22,7 @@ bool gtTextureD3D11::init( gtImage* image ){
 
 	if( !image ) return false;
 
-	bool is_RTT = (image->frames == -1);
+	bool is_RTT = (image->frames == (u32)-1);
 
 	if( !is_RTT )
 		image->convert( gtImageFormat::R8G8B8A8 );
